Relocated calculate_battery_level
